feat: 폴더 라벨 추가
