+++ components/icons/FolderIcon.tsx
@@ -1,14 +1,17 @@
 import * as React from 'react';
 import { SVGProps } from 'react';
-const SvgFolderIcon = (props: SVGProps<SVGSVGElement>) => (
-  <svg width={32} height={32} viewBox="0 0 32 32" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
-    <path
-      d="M27.113 26H4.925A.925.925 0 0 1 4 25.075V10h23a1 1 0 0 1 1 1v14.113a.888.888 0 0 1-.887.887ZM4 10V7a1 1 0 0 1 1-1h6.588a.987.987 0 0 1 .7.288L16 10"
-      stroke="#FF541E"
-      strokeWidth={2.311}
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  </svg>
-);
+const SvgFolderIcon = (props: SVGProps<SVGSVGElement>) => {
+  const { width, height } = props;
+  return (
+    <svg width={width} height={height} viewBox="0 0 32 32" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+      <path
+        d="M27.113 26H4.925A.925.925 0 0 1 4 25.075V10h23a1 1 0 0 1 1 1v14.113a.888.888 0 0 1-.887.887ZM4 10V7a1 1 0 0 1 1-1h6.588a.987.987 0 0 1 .7.288L16 10"
+        stroke="#FF541E"
+        strokeWidth={2.311}
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
 export default SvgFolderIcon;
